ml-server/requirements.txt
flask
flask-cors
opencv-python
numpy
torch
ultralytics
tensorflow
pillow
pytesseract
easyocr
